+++ plugin.php
@@ -11,8 +11,8 @@ License: GPLv2 or later
 License URI: https://www.gnu.org/licenses/gpl-2.0.html
 Text Domain: tasty-fonts
 Domain Path: /languages
-Requires at least: 6.1
-Tested up to: 6.9.4
+Requires at least: 6.5
+Tested up to: 6.8
 Requires PHP: 8.1
 */